[REMANIEMENT] Modifie la structure des fichiers concernant la sidebar
