+++ mon-aide-cyber-ui/src/composants/layout/layout-aidant/MenuNavigationElement.ts
@@ -1,0 +1,6 @@
+import { LienNavigation } from '../LayoutPublic';
+
+export type MenuNavigationElement = LienNavigation & {
+  actif: boolean;
+  enfants?: MenuNavigationElement[];
+};
