+++ mon-aide-cyber-ui/src/composants/layout/layout-aidant/sidebar/menu-navigation/MenuNavigation.tsx
@@ -1,0 +1,16 @@
+import { MenuNavigationElement } from '../../MenuNavigationElement';
+import { LienMenuNavigation } from './LienMenuNavigation';
+
+type ProprietesMenuNavigation = {
+  elements: MenuNavigationElement[];
+};
+
+export const MenuNavigation = ({ elements }: ProprietesMenuNavigation) => {
+  return (
+    <ul className="menu-lateral">
+      {elements?.map((element) => (
+        <LienMenuNavigation key={element.nom} element={element} />
+      ))}
+    </ul>
+  );
+};
